requirements.txt
absl-py>=2.1.0
appnope>=0.1.4
asttokens>=2.4.1
astunparse>=1.6.3
certifi>=2024.8.30
charset-normalizer>=3.4.0
comm>=0.2.2
contourpy>=1.3.0
cycler>=0.12.1
debugpy>=1.8.8
decorator>=5.1.1
executing>=2.1.0
flask>=3.0.2
flatbuffers>=24.3.25
fonttools>=4.54.1
gast>=0.6.0
google-pasta>=0.2.0
grpcio>=1.67.1
h5py>=3.12.1
idna>=3.10
iniconfig>=2.0.0
ipykernel>=6.29.5
ipython>=8.18.1
jedi>=0.19.2
joblib>=1.4.2
jupyter_client>=8.6.3
jupyter_core>=5.7.2
keras>=3.6.0
kiwisolver>=1.4.7
libclang>=18.1.1
Markdown>=3.7
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
matplotlib>=3.9.2
matplotlib-inline>=0.1.7
mdurl>=0.1.2
ml-dtypes>=0.4.1
namex>=0.0.8
nest-asyncio>=1.6.0
numpy>=1.26.4
opencv-python>=4.10.0.84
opt_einsum>=3.4.0
optree>=0.13.1
packaging>=24.2
parso>=0.8.4
path>=17.0.0
pexpect>=4.9.0
pillow>=9.0.0,<11.0.0
platformdirs>=4.3.6
pluggy>=1.5.0
prompt_toolkit>=3.0.48
protobuf>=4.25.5
psutil>=6.1.0
ptyprocess>=0.7.0
pure_eval>=0.2.3
Pygments>=2.18.0
pyparsing>=3.2.0
pytest>=8.3.3
python-dateutil>=2.9.0.post0
pyzmq>=26.2.0
requests>=2.32.3
rich>=13.9.4
scikit-learn>=1.5.2
scipy>=1.13.1
setuptools>=75.4.0
six>=1.16.0
stack-data>=0.6.3
streamlit>=1.32.0
tensorboard>=2.18.0
tensorboard-data-server>=0.7.2
tensorflow>=2.18.0
termcolor>=2.5.0
threadpoolctl>=3.5.0
tornado>=6.4.1
traitlets>=5.14.3
typing_extensions>=4.12.2
urllib3>=2.2.3
wcwidth>=0.2.13
Werkzeug>=3.1.3
wheel>=0.45.0
wrapt>=1.16.0